Added texture support to the material class
Former-commit-id: 55a99a1b9b6ef509d33e7e2443c944e5ae7ca96d [formerly 13311137682c2fe91516ef0e8a3abec3d8cc53a8]
Former-commit-id: fdf8dd7b4ba3785595081573dbf2cb7863bb76af

+++ GLEngineProject/include/engine/MaterialParameter.hpp
@@ -13,6 +13,6 @@ namespace engine {
 		public:
 			std::vector<GLubyte> data;
 			bool use = false;
-			void (ShaderProgram::*func)(const size_t &index, std::vector<GLubyte> data);
+			void (ShaderProgram::*func)(const unsigned int &index, std::vector<GLubyte> data);
 	};
 }
